requirements.txt
pyqt
pandas
yaml

